+++ src/ai/backend/kernel/r/__init__.py
@@ -25,8 +25,8 @@ class Runner(BaseRunner):
     }
     jupyter_kspec_name = 'ir'
 
-    def __init__(self):
-        super().__init__()
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
         self.input_queue = None
         self.output_queue = None